+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 4.0.3
+
+* refactor for AWS lambda support
+
 # 4.0.2
 
 * changed the methods for the suppression creation routes from `PATCH` > `POST`
