--- TODO.md
@@ -1,25 +1,0 @@
-# todo
-- [x] write examples GET
-- [ ] write tests GET
-- [ ] write tests POST / PATCH
-
-- [x] add all routes `POST` `PATCH`
-- [x] add all routes `DELETE`
-
-#### CREATE / UPDATE / DELETE
-```javascript
-sdk().createTeam(payload, callback)
-sdk().updateTeam(4, payload, callback)
-sdk().destroyTeam(4, payload, callback)
-```
-
-# todo publish
-- [ ] publish to evident module
-- [ ] deprecate esp_sdk.js module
-- [x] change default routes to 'esp.evident.io'
-- [ ] bump version number to 4
-
-
-# tests to write
-## live
-- [ ] createContactRequest
