+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "evident",
-  "version": "4.0.2",
+  "version": "4.0.3",
   "description": "Javascript SDK for Evident Security Platform - evident.io",
   "main": "index.js",
   "scripts": {
@@ -25,7 +25,7 @@
     "grunt-contrib-watch": "^0.6.1",
     "grunt-run": "^0.5.2",
     "grunt-standard": "^2.0.0",
-    "standard": "^5.4.1",
+    "standard": "^7.1.0",
     "tape": "^4.2.2"
   },
   "keywords": [
