+++ test/live/test_createSuborganization.js
@@ -1,4 +1,4 @@
-var sdk = require('../../index.js')
+var sdk = require('../../esp_sdk.js')
 
 // create team
 var payload = {
