+++ test/live/test_getDashboard.js
@@ -3,7 +3,7 @@ var test = require('tape')
 test('getDashboard', function (t) {
   t.plan(2)
 
-  var sdk = require('../../index.js')
+  var sdk = require('../../esp_sdk.js')
 
   var esp = sdk()
   esp.getDashboard(function (err, data) {
