+++ test/live/test_updateCustomSignature.js
@@ -1,4 +1,4 @@
-var sdk = require('../../index.js')
+var sdk = require('../../esp_sdk.js')
 var test = require('tape')
 
 test('updateCustomSignature', function (t) {
